use renamed module
