+++ llama_index/llama_pack/download.py
@@ -1,6 +1,6 @@
 from typing import Type
 
-from llama_index.download.download_utils import (
+from llama_index.download.module import (
     LLAMA_HUB_URL,
     MODULE_TYPE,
     download_llama_module,
